third_party/WebKit/LayoutTests/external/wpt/mediacapture-streams/MediaStreamTrack-MediaElement-disabled-video-is-black.https-expected.txt

This is a testharness.js-based test.
FAIL Tests that a disabled video track in a MediaStream is rendered as blackness assert_equals: No red component in pixel #0 expected 0 but got 255
Harness: the test ran to completion.
